+++ routers/ordenes_compra.py
@@ -1,5 +1,5 @@
 from fastapi import APIRouter
-from connection import get_connection_siesa
+from db.connection import get_connection_siesa
 from fastapi import HTTPException
 from utils.fetch_all import fetch_all
 from fastapi.responses import JSONResponse
